// frontend/src/generated/model/createActivityRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * AnthoLume API v1
 * REST API for AnthoLume document management system
 * OpenAPI spec version: 1.0.0
 */
import type { CreateActivityItem } from './createActivityItem';

export interface CreateActivityRequest {
  device_id: string;
  device_name: string;
  activity: CreateActivityItem[];
}